requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.6
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
praw==7.8.1
prawcore==2.4.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.17.0
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
vaderSentiment==3.3.2
websocket-client==1.8.0
